keyboard as context for cross-render context
