+++ client/packages/common/src/hooks/useKeyboard/index.ts
@@ -1,1 +1,1 @@
-export * from './useKeyboard';
+export * from './Keyboard';
